fix(review): move _summarize_streams to _execution_helpers to fix line limit
tools_execution.py exceeded the 1600-line exemption after rebase.
Move _summarize_streams (and its CaptureReadError/summarize_capture deps)
to _execution_helpers.py where _process_capture_stream already lives.
Update test monkeypatch targets accordingly.
